+++ provider-mobfox/src/main/java/com/deltadna/android/sdk/ads/provider/mobfox/MobFoxEventForwarder.java
@@ -16,16 +16,22 @@
 
 package com.deltadna.android.sdk.ads.provider.mobfox;
 
+import android.util.Log;
+
 import com.deltadna.android.sdk.ads.bindings.AdRequestResult;
 import com.deltadna.android.sdk.ads.bindings.MediationAdapter;
 import com.deltadna.android.sdk.ads.bindings.MediationListener;
 import com.mobfox.sdk.interstitialads.InterstitialAd;
 import com.mobfox.sdk.interstitialads.InterstitialAdListener;
 
+import java.util.Locale;
+
 final class MobFoxEventForwarder implements InterstitialAdListener {
     
     private final MediationListener listener;
     private final MediationAdapter adapter;
+    
+    private boolean completed;
     
     MobFoxEventForwarder(MediationListener listener, MediationAdapter adapter) {
         this.listener = listener;
@@ -34,11 +40,18 @@ final class MobFoxEventForwarder implements InterstitialAdListener {
     
     @Override
     public void onInterstitialLoaded(InterstitialAd ad) {
+        Log.d(BuildConfig.LOG_TAG, "Interstitial loaded: " + ad);
         listener.onAdLoaded(adapter);
     }
     
     @Override
     public void onInterstitialFailed(InterstitialAd ad, Exception e) {
+        Log.w(BuildConfig.LOG_TAG, String.format(
+                Locale.US,
+                "Interstitial failed: %s/%s",
+                ad,
+                e));
+        
         // messages taken from decompiled InterstitialAd class
         final AdRequestResult result;
         switch (e.getMessage()) {
@@ -60,21 +73,25 @@ final class MobFoxEventForwarder implements InterstitialAdListener {
     
     @Override
     public void onInterstitialShown(InterstitialAd ad) {
+        Log.d(BuildConfig.LOG_TAG, "Interstitial shown: " + ad);
         listener.onAdShowing(adapter);
     }
     
     @Override
     public void onInterstitialClicked(InterstitialAd ad) {
+        Log.d(BuildConfig.LOG_TAG, "Interstitial clicked: " + ad);
         listener.onAdClicked(adapter);
     }
     
     @Override
-    public void onInterstitialClosed(InterstitialAd ad) {
-        listener.onAdClosed(adapter, false); // TODO check
+    public void onInterstitialFinished() {
+        Log.d(BuildConfig.LOG_TAG, "Interstitial finished");
+        completed = true;
     }
     
     @Override
-    public void onInterstitialFinished() {
-        listener.onAdClosed(adapter, true);
+    public void onInterstitialClosed(InterstitialAd ad) {
+        Log.d(BuildConfig.LOG_TAG, "Interstitial closed: " + ad);
+        listener.onAdClosed(adapter, completed);
     }
 }