+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.6.0](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.6.0) (YYYY-MM-DD)
 Added rewarded ads for Facebook.  
+Updated build and runtime dependencies.  
 
 ## [1.5.3](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.5.3) (2017-09-20)
 Workaround for hardware not supported error when using Tapjoy on Unity.