+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.6.0](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.6.0) (YYYY-MM-DD)
 Added rewarded ads for Facebook.  
+Fixed MobFox completed state.  
 Updated AdMob ads.  
 Updated AppLovin ads.  
 Updated Chartboost ads.  
@@ -9,6 +10,7 @@ Updated Facebook ads.
 Updated Flurry ads.  
 Updated InMobi ads.  
 Updated IronSource ads.  
+Updated MobFox ads.  
 Updated build and runtime dependencies.  
 
 ## [1.5.3](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.5.3) (2017-09-20)