+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updated InMobi ads.
 Updated IronSource ads.  
 Updated MobFox ads.  
 Updated Unity ads.  
+Updated Vungle ads.  
 Updated build and runtime dependencies.  
 
 ## [1.5.3](https://github.com/deltaDNA/android-smartads-sdk/releases/tag/1.5.3) (2017-09-20)